feat: refactor shared package structure and consolidate message types
  - Move web-only interfaces (ClientState, AppState, ServerState) from shared to web package
  - Consolidate message-types into single organized file with grouped constants
  - Replace individual message constants with grouped objects (HandshakeMessages, AppManagementMessages, etc.)
  - Update all imports across web and socket packages to use new grouped constants
  - Improve package boundaries and reduce coupling between packages

  Breaking Changes:
  - Message constants now accessed via grouped objects (e.g., HandshakeMessages.DA_HELLO vs DA_HELLO)
  - Web-specific state interfaces moved from @finos/fdc3-sail-shared to local types
