--- packages/common/src/app-hosting.ts
@@ -1,8 +1,0 @@
-
-export enum AppHosting { Frame, Tab }
-
-
-
-export interface SailHostManifest {
-    forceNewWindow: boolean
-}
