+++ packages/example-apps/tsconfig.json
@@ -24,7 +24,7 @@
   ],
   "references": [
     {
-      "path": "../common"
+      "path": "../shared"
     }
   ]
 }
