+++ packages/shared/src/constants.ts
@@ -1,0 +1,1 @@
+export enum AppHosting { Frame, Tab }
