+++ packages/shared/src/index.ts
@@ -1,0 +1,3 @@
+// Shared constants and message types for inter-package communication
+export * from "./constants"
+export * from "./message-types"
