+++ packages/shared/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../tsconfig.root.json",
   "include": [
-    "src/*"
+    "src/**/*"
   ],
   "compilerOptions": {
     "outDir": "dist",
