+++ packages/socket/src/__tests__/SailFDC3Server.test.ts
@@ -4,7 +4,7 @@ import { Socket } from "socket.io"
 import { SailFDC3Server, mapChannels } from "../desktop-agent/sailFDC3Server"
 import { SailAppInstanceManager } from "../desktop-agent/sailAppInstanceManager"
 import { AppDirectoryManager } from "../app-directory/appDirectoryManager"
-import { DesktopAgentHelloArgs, TabDetail } from "@finos/fdc3-sail-common"
+import { DesktopAgentHelloArgs, TabDetail } from "@finos/fdc3-sail-shared"
 import { ChannelType } from "@finos/fdc3-web-impl"
 import { getTestServer, clearSessions } from "./setup/setupTests"
 import path from "path"
