+++ packages/web/package.json
@@ -25,7 +25,7 @@
     "vite": "^5.0.2"
   },
   "dependencies": {
-    "@finos/fdc3-sail-common": "0.0.0",
+    "@finos/fdc3-sail-shared": "0.0.0",
     "@finos/fdc3-web-impl": "2.2.0-beta.4",
     "@types/react-dom": "^18.2.25",
     "@types/uuid": "^9.0.8",
