+++ packages/web/src/app/channel.tsx
@@ -1,4 +1,4 @@
-import { TabDetail } from "@finos/fdc3-sail-common"
+import { TabDetail } from "@finos/fdc3-sail-shared"
 import { Icon } from "./Icon"
 import styles from "./styles.module.css"
