+++ packages/web/src/app/intent-resolver.tsx
@@ -5,7 +5,7 @@ import {
   isFdc3UserInterfaceHandshake,
   isFdc3UserInterfaceResolve,
 } from "@finos/fdc3-schema/dist/generated/api/BrowserTypes"
-import { AugmentedAppIntent } from "@finos/fdc3-sail-common"
+import { AugmentedAppIntent } from "@finos/fdc3-sail-shared"
 import { channels, handleChannelUpdates, setAppChannel } from "./util"
 type IframeResolveAction = BrowserTypes.Fdc3UserInterfaceResolveAction
 type IframeResolvePayload = BrowserTypes.Fdc3UserInterfaceResolvePayload
