+++ packages/web/src/app/util.ts
@@ -1,6 +1,6 @@
 import { InstanceID } from "@finos/fdc3-web-impl"
 import { io, Socket } from "socket.io-client"
-import { CHANNEL_RECEIVER_HELLO, CHANNEL_RECEIVER_UPDATE, ChannelReceiverHelloRequest, ChannelReceiverUpdate, FDC3_APP_EVENT, FDC3_DA_EVENT, SAIL_INTENT_RESOLVE_ON_CHANNEL, SailIntentResolveOpenChannelArgs, TabDetail } from "@finos/fdc3-sail-common"
+import { ChannelMessages, AppManagementMessages, IntentMessages, ChannelReceiverHelloRequest, ChannelReceiverUpdate, SailIntentResolveOpenChannelArgs, TabDetail } from "@finos/fdc3-sail-shared"
 
 
 export const channels: TabDetail[] = []
@@ -12,7 +12,7 @@ export function handleChannelUpdates(renderChannels: () => void) {
             userSessionId: getUserSessionId(),
             instanceId: getInstanceId(),
         }
-        const result: ChannelReceiverUpdate | undefined = await socket.emitWithAck(CHANNEL_RECEIVER_HELLO, msg)
+        const result: ChannelReceiverUpdate | undefined = await socket.emitWithAck(ChannelMessages.CHANNEL_RECEIVER_HELLO, msg)
         if (result) {
             channels.length = 0;
             channels.push(...result.tabs)
@@ -20,7 +20,7 @@ export function handleChannelUpdates(renderChannels: () => void) {
         }
     })
 
-    socket.on(CHANNEL_RECEIVER_UPDATE, (data: ChannelReceiverUpdate) => {
+    socket.on(ChannelMessages.CHANNEL_RECEIVER_UPDATE, (data: ChannelReceiverUpdate) => {
         channels.length = 0;
         channels.push(...data.tabs)
         renderChannels()
@@ -34,20 +34,20 @@ export async function setAppChannel(appId: string, channel: string): Promise<voi
         channel
     }
 
-    await socket.emitWithAck(SAIL_INTENT_RESOLVE_ON_CHANNEL, msg)
+    await socket.emitWithAck(IntentMessages.SAIL_INTENT_RESOLVE_ON_CHANNEL, msg)
     console.log("SET APP CHANNEL: " + JSON.stringify(msg))
 }
 
 /* eslint-disable  @typescript-eslint/no-explicit-any */
 export function link(socket: Socket, channel: MessageChannel, source: InstanceID) {
-    socket.on(FDC3_DA_EVENT, (data: any) => {
+    socket.on(AppManagementMessages.FDC3_DA_EVENT, (data: any) => {
         // console.log(`DA Sent ${JSON.stringify(data)} from socket`)
         channel.port2.postMessage(data)
     })
 
     channel.port2.onmessage = function (event) {
         // console.log(`App Sent ${JSON.stringify(event.data)} from message port`)
-        socket.emit(FDC3_APP_EVENT, event.data, source)
+        socket.emit(AppManagementMessages.FDC3_APP_EVENT, event.data, source)
     }
 }
