+++ packages/web/tsconfig.json
@@ -24,7 +24,7 @@
   ],
   "references": [
     {
-      "path": "../common"
+      "path": "../shared"
     }
   ]
 }
